finally modularize everything
